requirements.txt
openai==2.29.0
feedparser==6.0.12
requests==2.32.3
numpy==1.26.4
scikit-learn==1.2.2
python-dotenv==1.0.1
fastapi==0.115.11
uvicorn==0.34.0
jinja2==3.1.3
apscheduler==3.11.2
